requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
geographiclib==2.0
geopy==2.4.1
gpxpy==1.6.2
idna==3.7
iniconfig==2.0.0
numpy==1.24.4
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
rdp==0.8
requests==2.32.3
SRTM.py==0.3.7
tomli==2.0.1
urllib3==2.2.1
